refactor(forge): split into neutral core + github backend behind forge dispatch
